# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inner Circle Partners Portal API",
    "description": "API documentation for the Inner Circle Partners Portal - a comprehensive partner referral management system.\n\n## Authentication\n\nMost endpoints require authentication via Clerk. Include the session token in your requests.\n\n## Rate Limiting\n\nAPI requests are rate-limited to prevent abuse:\n- General endpoints: 100 requests per minute\n- Auth endpoints: 10 requests per minute\n- Webhook endpoints: 1000 requests per minute\n\n## Response Format\n\nAll responses follow a standard format:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"meta\": { \"timestamp\": \"...\" }\n}\n```\n\nError responses:\n```json\n{\n  \"success\": false,\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human readable message\"\n  }\n}\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Inner Circle Support",
      "email": "support@innercircle.co"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://innercircle.co/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.innercircle.co",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/partners/{id}/stats": {
      "get": {
        "tags": ["Partners"],
        "summary": "Get partner statistics",
        "description": "Returns detailed statistics for a partner including earnings, referrals, and conversion rates",
        "operationId": "getPartnerStats",
        "security": [{ "clerkAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Partner ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partner statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerStatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/partners/interest": {
      "post": {
        "tags": ["Partners"],
        "summary": "Submit partner interest",
        "description": "Submit an application to become a partner",
        "operationId": "submitInterest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInterestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interest submitted"
          },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/payments/connect": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create Stripe Connect account",
        "description": "Creates a Stripe Connect Express account for the partner",
        "operationId": "createConnectAccount",
        "security": [{ "clerkAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["partnerId"],
                "properties": {
                  "partnerId": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "country": { "type": "string", "default": "US" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account created with onboarding URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accountId": { "type": "string" },
                        "onboardingUrl": { "type": "string", "format": "uri" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payments/dashboard": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get Stripe dashboard link",
        "description": "Returns a link to the partner's Stripe Express dashboard",
        "operationId": "getDashboardLink",
        "security": [{ "clerkAuth": [] }],
        "parameters": [
          {
            "name": "partnerId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "url": { "type": "string", "format": "uri" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      },
      "adminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Requires admin role in Clerk user metadata"
      }
    },
    "schemas": {
...
      "PartnerStatsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "totalEarnings": { "type": "integer", "description": "Total earnings in cents" },
              "pendingEarnings": { "type": "integer" },
              "totalReferrals": { "type": "integer" },
              "conversionRate": { "type": "number" },
              "activeCampaigns": { "type": "integer" },
              "tier": { "type": "string" },
              "nextTierProgress": { "type": "number" }
            }
          }
        }
      },
...
      "PartnerInterestRequest": {
        "type": "object",
        "required": ["email", "name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "company": { "type": "string" },
          "website": { "type": "string", "format": "uri" },
          "message": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "FORBIDDEN",
                "message": "You do not have permission to access this resource"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "NOT_FOUND",
                "message": "The requested resource was not found"
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid request data",
                "details": {
                  "errors": {
                    "email": ["Invalid email format"]
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}